+++ tests/vector_stores/test_postgres.py
@@ -145,7 +145,7 @@ def hybrid_node_embeddings() -> List[NodeWithEmbedding]:
         NodeWithEmbedding(
             embedding=_get_sample_vector(5.0),
             node=TextNode(
-                text="The quick brown fox jumped over the lazy dog (fox).",
+                text="The quick brown fox jumped over the lazy dog.",
                 id_="ccc",
                 relationships={NodeRelationship.SOURCE: RelatedNodeInfo(node_id="ccc")},
             ),
@@ -176,7 +176,7 @@ async def test_instance_creation(db: None) -> None:
 
 @pytest.mark.skipif(postgres_not_available, reason="postgres db is not available")
 @pytest.mark.asyncio
-@pytest.mark.parametrize("use_async", [(True,), (False,)])
+@pytest.mark.parametrize("use_async", [True, False])
 async def test_add_to_db_and_query(
     pg: PGVectorStore, node_embeddings: List[NodeWithEmbedding], use_async: bool
 ) -> None:
@@ -197,7 +197,7 @@ async def test_add_to_db_and_query(
 
 @pytest.mark.skipif(postgres_not_available, reason="postgres db is not available")
 @pytest.mark.asyncio
-@pytest.mark.parametrize("use_async", [(True,), (False,)])
+@pytest.mark.parametrize("use_async", [True, False])
 async def test_add_to_db_and_query_with_metadata_filters(
     pg: PGVectorStore, node_embeddings: List[NodeWithEmbedding], use_async: bool
 ) -> None:
@@ -223,7 +223,7 @@ async def test_add_to_db_and_query_with_metadata_filters(
 
 @pytest.mark.skipif(postgres_not_available, reason="postgres db is not available")
 @pytest.mark.asyncio
-@pytest.mark.parametrize("use_async", [(True,), (False,)])
+@pytest.mark.parametrize("use_async", [True, False])
 async def test_add_to_db_query_and_delete(
     pg: PGVectorStore, node_embeddings: List[NodeWithEmbedding], use_async: bool
 ) -> None:
@@ -255,7 +255,39 @@ async def test_add_to_db_query_and_delete(
 
 @pytest.mark.skipif(postgres_not_available, reason="postgres db is not available")
 @pytest.mark.asyncio
-@pytest.mark.parametrize("use_async", [(True,), (False,)])
+@pytest.mark.parametrize("use_async", [True, False])
+async def test_sparse_query(
+    pg_hybrid: PGVectorStore,
+    hybrid_node_embeddings: List[NodeWithEmbedding],
+    use_async: bool,
+) -> None:
+    if use_async:
+        await pg_hybrid.async_add(hybrid_node_embeddings)
+    else:
+        pg_hybrid.add(hybrid_node_embeddings)
+    assert isinstance(pg_hybrid, PGVectorStore)
+
+    # text search should work when query is a sentence and not just a single word
+    q = VectorStoreQuery(
+        query_embedding=_get_sample_vector(0.1),
+        query_str="who is the fox?",
+        sparse_top_k=2,
+        mode=VectorStoreQueryMode.SPARSE,
+    )
+
+    if use_async:
+        res = await pg_hybrid.aquery(q)
+    else:
+        res = pg_hybrid.query(q)
+    assert res.nodes
+    assert len(res.nodes) == 2
+    assert res.nodes[0].node_id == "ccc"
+    assert res.nodes[1].node_id == "ddd"
+
+
+@pytest.mark.skipif(postgres_not_available, reason="postgres db is not available")
+@pytest.mark.asyncio
+@pytest.mark.parametrize("use_async", [True, False])
 async def test_hybrid_query(
     pg_hybrid: PGVectorStore,
     hybrid_node_embeddings: List[NodeWithEmbedding],
@@ -304,10 +336,29 @@ async def test_hybrid_query(
     assert res.nodes[2].node_id == "ccc"
     assert res.nodes[3].node_id == "ddd"
 
+    # text search should work when query is a sentence and not just a single word
+    q = VectorStoreQuery(
+        query_embedding=_get_sample_vector(0.1),
+        query_str="who is the fox?",
+        similarity_top_k=2,
+        mode=VectorStoreQueryMode.HYBRID,
+    )
+
+    if use_async:
+        res = await pg_hybrid.aquery(q)
+    else:
+        res = pg_hybrid.query(q)
+    assert res.nodes
+    assert len(res.nodes) == 4
+    assert res.nodes[0].node_id == "aaa"
+    assert res.nodes[1].node_id == "bbb"
+    assert res.nodes[2].node_id == "ccc"
+    assert res.nodes[3].node_id == "ddd"
+
 
 @pytest.mark.skipif(postgres_not_available, reason="postgres db is not available")
 @pytest.mark.asyncio
-@pytest.mark.parametrize("use_async", [(True,), (False,)])
+@pytest.mark.parametrize("use_async", [True, False])
 async def test_add_to_db_and_hybrid_query_with_metadata_filters(
     pg_hybrid: PGVectorStore,
     hybrid_node_embeddings: List[NodeWithEmbedding],